requirements.txt
websocket~=0.2.1
websocket-client~=1.2.3
discord~=1.7.3
python-binance
